refactor: Move streaming source settings to Player & Audio, keep Tidal login in Integration
- New StreamingSourcesSettings screen (settings/player/sources) holds all source
  settings: audio source ordering, Tidal enable/account-first/quality, Tidal
  instances (add/check/discover/reset), Tidal artwork, Deezer, Amazon
- TidalSettings (Integration → Tidal) is now login/logout only, with a note
  pointing to the moved settings
- PlayerSettings links to the new Streaming sources screen
- Registered settings/player/sources route in NavigationBuilder
- Aligned enable-by-default values across UI + resolver (Tidal/Deezer on, Amazon off)
  in sourceResolutionChain() and tidalSourceApplies()
- Added streaming_sources / streaming_sources_description / tidal_source_settings_moved strings
