+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/screens/NavigationBuilder.kt
@@ -62,6 +62,7 @@ import moe.rukamori.archivetune.ui.screens.settings.HiddenPlaylistsScreen
 import moe.rukamori.archivetune.ui.screens.settings.IconScreen
 import moe.rukamori.archivetune.ui.screens.settings.IntegrationScreen
 import moe.rukamori.archivetune.ui.screens.settings.InternetSettings
+import moe.rukamori.archivetune.ui.screens.settings.StreamingSourcesSettings
 import moe.rukamori.archivetune.ui.screens.settings.TidalSettings
 import moe.rukamori.archivetune.ui.screens.settings.LastFMSettings
 import moe.rukamori.archivetune.ui.screens.settings.LogcatScreen
@@ -406,6 +407,9 @@ fun NavGraphBuilder.navigationBuilder(
     }
     composable("settings/player/chiper") {
         ChiperSettings(navController)
+    }
+    composable("settings/player/sources") {
+        StreamingSourcesSettings(navController)
     }
     composable("settings/storage") {
         StorageSettings(navController)
